Fix bug that causes LHAPDF to crash
This bug was discovered when trying to use T2KReWeight on GENIE files. It seems that the directory that was opened was never closed, so at one point the total number of "open directories" exceeded the current limit for gSystem (bash: ulimit -S -n), causing LHAPDF to crash. If the directory is freed every time we pass that line, the error does not occur. Thanks to Jacek M. Holeczek for proposing this solution!